+++ CHANGELOG.md
@@ -11,13 +11,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Fixed
 
+## [0.7.0] - 2021-10-03
+### Added 
+- Add travel arrangements page
+  - Allow ADMIN to insert and see all travel arrangements
+  - Allow TRAVEL GUIDE only to see travel arrangements
+
 ## [0.6.1] - 2021-10-03
 ### Fixed
 - When a user registering in the app first time, redirect them to the index(home) page
 - If there is no action related to changing account type request, do not display desired account type
 - Show a message related to waiting for a request (related to the change of the account type) to be resolved
 only for a user who has sent the request.
-- Set LOGIN_DISABLED to be False when we load development environment configuration  
+- Set LOGIN_DISABLED to be False when we load development environment configuration
+
 ## [0.6.0] - 2021-10-03
 ### Added 
 - Edit user data page